# contracts/openapi/ttndd-ops-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/hrm/members": {
      "post": {
        "operationId": "HrmController_createMember",
        "summary": "Create a new member with profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "HrmController_findMany",
        "summary": "List members with filters (paginated)",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}": {
      "get": {
        "operationId": "HrmController_findById",
        "summary": "Get member detail with profile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "HrmController_updateProfile",
        "summary": "Update member profile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}/privacy-export": {
      "get": {
        "operationId": "HrmController_exportPersonalData",
        "summary": "Export personal data for data-subject access requests",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}/transition": {
      "post": {
        "operationId": "HrmController_transitionStatus",
        "summary": "Transition member status (state machine SM-1)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}/transfer": {
      "post": {
        "operationId": "HrmController_transferMember",
        "summary": "Transfer member to different branch/unit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}/assign-unit": {
      "post": {
        "operationId": "HrmController_assignUnit",
        "summary": "T-1007: Assign member to a unit within their branch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/org-chart": {
      "get": {
        "operationId": "HrmController_getOrgChart",
        "summary": "Get organization chart (flat list)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/org-chart/tree": {
      "get": {
        "operationId": "HrmController_getOrgChartTree",
        "summary": "T-1006: Get hierarchical org chart tree with member counts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}/timeline": {
      "get": {
        "operationId": "HrmController_getTimeline",
        "summary": "T-1009: Get member event timeline (domain events + branch history)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/transfers/{transferId}/sign": {
      "post": {
        "operationId": "HrmController_signTransfer",
        "summary": "T-1010: Sign a transfer handover record",
        "parameters": [
          {
            "name": "transferId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignTransferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/stats": {
      "get": {
        "operationId": "HrmController_getStats",
        "summary": "Get HRM statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}/character-sheet": {
      "get": {
        "operationId": "HrmController_getCharacterSheet",
        "summary": "T-0048: Get cross-module character sheet (profile + rewards + rank + attendance)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/members/{id}/compliance": {
      "get": {
        "operationId": "HrmController_checkCompliance",
        "summary": "T-0050/T-1005: Check member compliance (guardian, medical, background check)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/hrm/compliance/dashboard": {
      "get": {
        "operationId": "HrmController_getComplianceDashboard",
        "summary": "T-1005: Org-wide compliance dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["HRM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TTNDD_OPS API",
    "description": "Thanh Thiếu Niên Đại Đạo — Platform REST API\n\nMulti-tenant ERP platform for DTNDD scouting operations.\nAll endpoints require Bearer token and X-Org-Id header for RLS isolation.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateMemberProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn An"
          },
          "birthDate": {
            "type": "string",
            "example": "2012-08-22"
          },
          "gender": {
            "type": "string",
            "example": "male"
          },
          "address": {
            "type": "string",
            "example": "Q.1, TP. HCM"
          },
          "personalPhone": {
            "type": "string",
            "example": "0901234567"
          },
          "personalEmail": {
            "type": "string",
            "example": "an@email.com"
          },
          "guardianName": {
            "type": "string",
            "example": "Trần Thị Mẹ"
          },
          "guardianPhone": {
            "type": "string",
            "example": "0909876543"
          },
          "guardianRelation": {
            "type": "string",
            "example": "mother"
          },
          "healthNotes": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string",
            "example": "0901234999"
          }
        },
        "required": ["fullName"]
      },
      "CreateMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "user-uuid-here"
          },
          "role": {
            "type": "string",
            "example": "member",
            "enum": ["member", "leader", "admin"]
          },
          "branchId": {
            "type": "string"
          },
          "unitId": {
            "type": "string"
          },
          "memberCode": {
            "type": "string",
            "example": "DS-001"
          },
          "scoutName": {
            "type": "string",
            "example": "Sơn Ca"
          },
          "heroName": {
            "type": "string",
            "example": "Chiến binh Ánh Sáng"
          },
          "profile": {
            "$ref": "#/components/schemas/CreateMemberProfileDto"
          }
        },
        "required": ["userId", "role", "profile"]
      },
      "UpdateMemberProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "personalPhone": {
            "type": "string"
          },
          "personalEmail": {
            "type": "string"
          },
          "guardianName": {
            "type": "string"
          },
          "guardianPhone": {
            "type": "string"
          },
          "guardianRelation": {
            "type": "string"
          },
          "healthNotes": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string"
          }
        }
      },
      "TransitionStatusDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "example": "approve",
            "enum": [
              "approve",
              "reject",
              "voluntary_pause",
              "resume",
              "discipline",
              "reinstate",
              "branch_transition",
              "accept_in_new_branch",
              "offboarding"
            ]
          }
        },
        "required": ["action"]
      },
      "TransferMemberDto": {
        "type": "object",
        "properties": {
          "toBranchId": {
            "type": "string"
          },
          "toUnitId": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "example": "Chuyển do thay đổi nơi sinh hoạt"
          }
        },
        "required": ["toBranchId"]
      },
      "AssignUnitDto": {
        "type": "object",
        "properties": {
          "unitId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": ["unitId"]
      },
      "SignTransferDto": {
        "type": "object",
        "properties": {
          "signatureNote": {
            "type": "string",
            "example": "Xác nhận tiếp nhận đoàn sinh"
          },
          "accepted": {
            "type": "boolean"
          }
        },
        "required": ["signatureNote"]
      },
...
    }
  }
}